Move CardsExampleNode to its own package
